Rename test-runner back + rename Main classes so they are more predictable
