+++ runners/command-runner/src/main/scala/stryker4s/command/config/ProcessRunnerConfig.scala
@@ -2,4 +2,4 @@ package stryker4s.command.config
 
 import stryker4s.run.process.Command
 
-case class ProcessRunnerConfig(testRunnerCommand: Command)
+case class ProcessRunnerConfig(testRunner: Command)
