+++ runners/maven/src/main/scala/stryker4s/maven/Stryker4sMain.scala
@@ -1,4 +1,4 @@
-package stryker4s.run
+package stryker4s.maven
 
 import javax.inject.Inject
 import org.apache.maven.plugin.{AbstractMojo, MojoFailureException}
@@ -9,7 +9,7 @@ import stryker4s.run.threshold.ErrorStatus
 /** The main goal for this plugin. Starts Stryker4s.
   */
 @Mojo(name = "run")
-class Stryker4sMavenPlugin @Inject()(@Parameter(defaultValue = "${project}") project: MavenProject)
+class Stryker4sMain @Inject()(@Parameter(defaultValue = "${project}") project: MavenProject)
     extends AbstractMojo {
   override def execute(): Unit = {
     new Stryker4sMavenRunner(project).run() match {
