+++ runners/maven/src/main/scala/stryker4s/maven/Stryker4sMavenRunner.scala
@@ -1,10 +1,12 @@
-package stryker4s.run
+package stryker4s.maven
+
 import org.apache.maven.project.MavenProject
 import org.apache.maven.shared.invoker.{DefaultInvoker, Invoker}
 import stryker4s.config.Config
 import stryker4s.mutants.applymutants.ActiveMutationContext.{envVar, ActiveMutationContext}
 import stryker4s.mutants.findmutants.SourceCollector
 import stryker4s.report.Reporter
+import stryker4s.maven.runner.MavenMutantRunner
 
 class Stryker4sMavenRunner(project: MavenProject) extends Stryker4sRunner {
   override def resolveRunner(collector: SourceCollector, reporter: Reporter)(implicit config: Config): MutantRunner =
