+++ runners/sbt/src/main/scala/stryker4s/sbt/Stryker4sMain.scala
@@ -13,7 +13,7 @@ import scala.language.implicitConversions
 /**
   * This plugin adds a new task (stryker) to the project that allow you to run mutation testing over your code
   */
-object Stryker4sPlugin extends AutoPlugin {
+object Stryker4sMain extends AutoPlugin {
 
   override def requires = JvmPlugin
